have added a lot in this commit.
At this stage, all APIs are linked up, and when the user
searches for a city in the searchbar and clicks the search button,
all weather information is displayed on screen (current weather, and
the next 5 day weather, and the current UV Index.)

When a search is created, a button is created with the name of the
recent search. However, at this stage, when these buttons have been
clicked, nothing happens. Next objective is to save weather info
into these buttons